Implemented voting, but had to totally change the layout of the app
Since on creating fragments the rest of the code in the method still
runs we have make it so each button launches the next fragment. I'm not
sure if this is the most efficient way to do things but it certainly is
suuuper easy. now we just have to finish the president policy fragment
and make the chancellor policy fragment. Oh and we have to figure out
the listview in the chooseChancellorFragemnt.